+++ src/constants/action_constants.js
@@ -14,6 +14,7 @@ const actions = {
   GET_USER_CART_FAILURE: "GET_USER_CART_FAILURE",
 
   ADD_PRODUCT_TO_CART: "ADD_PRODUCT_TO_CART",
+  DELETE_PRODUCT_FROM_CART: " DELETE_PRODUCT_FROM_CART",
 
   GET_CATEGORY_NAME: "GET_CATEGORY_NAME",
   GET_CATEGORY_NAME_SUCCESS: "GET_CATEGORY_NAME_SUCCESS",
