+++ src/pages/Account/Components/UserPage/CartItem.jsx
@@ -1,8 +1,14 @@
 import { useState } from "react";
 import { FaGreaterThan, FaRegHeart, FaStar } from "react-icons/fa";
+import { useDispatch } from "react-redux";
+import { deleteProductFromCart } from "../../../../redux/actions";
 
 const CartItem = ({ data }) => {
+  const dsipatch = useDispatch();
   const [qty, setQty] = useState(data.qty);
+  const deleteHandler = (id) => {
+    dsipatch(deleteProductFromCart(id));
+  };
   return (
     <div className="body">
       <div className="item">
@@ -19,7 +25,10 @@ const CartItem = ({ data }) => {
               <FaRegHeart />
             </div>
             <div className="remove_product">
-              <button> Delete </button>
+              <button onClick={() => deleteHandler(data.product.id)}>
+                {" "}
+                Delete{" "}
+              </button>
             </div>
           </div>
         </div>
