+++ src/pages/Account/Components/UserPage/Wishlist.jsx
@@ -5,23 +5,7 @@ import { toast } from "react-toastify";
 import CartItemsHeader from "./CartItemsHeader";
 import WishlistCartItem from "./WishlistCartItem";
 
-const WishlistPage = () => {
-  const [wishlist, setWishlist] = useState([]);
-
-  const getWishlistProducts = async () => {
-    try {
-      const { data } = await axios.get(
-        "https://fakestoreapi.com/products?limit=20"
-      );
-      setWishlist(data);
-    } catch (error) {
-      toast.error("Can't get wishlist");
-    }
-  };
-
-  useEffect(() => {
-    getWishlistProducts();
-  }, []);
+const WishlistPage = ({ wishlist }) => {
   return (
     <div className="wishlist_cart_items">
       <CartItemsHeader />
